[fix] chr() needs an int value, not a float

+++ scripts/heatled.py
@@ -9,6 +9,6 @@ def limit(value, lower, upper):
 if __name__ == "__main__":
 	while True:
 		capted_temp = float(open(HALFS_ROOT+"/sensors/temp_radiator").read().strip("\0"))
-		transformed_temp = 255*capted_temp
+		transformed_temp = int(255*capted_temp)
 		open(HALFS_ROOT+"/animations/heater/frames").write(chr(limit(transformed_temp, 0, 255)))
 		sleep(1)
